Migrate UIImage+Helpers to Swift
