--- ios/keyman/Keyman/Keyman/Categories/UIImage+Helpers.m
@@ -1,32 +1,0 @@
-//
-//  UIImage+Helpers.m
-//  Keyman
-//
-//  Created by Serkan Kurt on 11/12/2014.
-//  Copyright (c) 2017 SIL International. All rights reserved.
-//
-
-#import "UIImage+Helpers.h"
-
-@implementation UIImage (Helpers)
-
-- (UIImage *)scaleToSize:(CGSize)newSize {
-    UIGraphicsBeginImageContextWithOptions(newSize, NO, 0.0);
-    [self drawInRect:CGRectMake(0, 0, newSize.width, newSize.height)];
-    UIImage *newImage = UIGraphicsGetImageFromCurrentImageContext();
-    UIGraphicsEndImageContext();
-    return newImage;
-}
-
-- (UIImage *)tintedImageUsingColor:(UIColor *)tintColor {
-    UIGraphicsBeginImageContextWithOptions(self.size, NO, self.scale);
-    CGRect drawRect = CGRectMake(0, 0, self.size.width, self.size.height);
-    [self drawInRect:drawRect];
-    [tintColor set];
-    UIRectFillUsingBlendMode(drawRect, kCGBlendModeSourceAtop);
-    UIImage *tintedImage = UIGraphicsGetImageFromCurrentImageContext();
-    UIGraphicsEndImageContext();
-    return tintedImage;
-}
-
-@end
